This was a bug fixing commit, bring the program up to a functional stage where the screen is drawn and the logo is loaded and appears on the screen correctly.
To do this I had to make a few small changes, such as changing the
parameters in updateScreen, importing files where necessary and changing
the ! syntax to 'not'.

I have also changed the background color to black to make it more
authentic.

The next commit will involve making the logo actually move around the
screen, which it currently fails to do.